Ast

Description

This module describes the Relapse's abstract syntax tree.

It also contains some simple functions for the map of references that a Relapse grammar consists of.

Finally it also contains some very simple pattern functions.

Documentation

data Pattern

Pattern recursively describes a Relapse Pattern.

Constructors

Empty

ZAny

Node (Expr Bool) Pattern

Or Pattern Pattern

And Pattern Pattern

Not Pattern

Concat Pattern Pattern

Interleave Pattern Pattern

ZeroOrMore Pattern

Optional Pattern

Contains Pattern

Reference String

data Grammar

Refs is a map from reference name to pattern and describes a relapse grammar.

emptyRef :: Grammar

emptyRef returns an empty reference map.

union :: Grammar -> Grammar -> Grammar

union returns the union of two reference maps.

newRef :: String -> Pattern -> Grammar

newRef returns a new reference map given a single pattern and its reference name.

reverseLookupRef :: Pattern -> Grammar -> Maybe String

reverseLookupRef returns the reference name for a given pattern.

lookupRef :: Grammar -> String -> Either String Pattern

lookupRef looks up a pattern in the reference map, given a reference name.

hasRecursion :: Grammar -> Either String Bool

hasRecursion returns whether an relapse grammar has any recursion, starting from the "main" reference.

listRefs :: Grammar -> [String]

listRefs returns the list of reference names.

nullable :: Grammar -> Pattern -> Either String Bool

The nullable function returns whether a pattern is nullable. This means that the pattern matches the empty string.