Data.Katydid.Relapse.Exprs.Compare

Description

This module contains the Relapse compare expressions: equal, not equal, greater than, greater than or equal, less than and less than or equal.

Documentation

mkEqExpr :: [AnyExpr] -> Either String AnyExpr

mkEqExpr dynamically creates an eq (equal) expression, if the two input types are the same.

eqExpr :: Eq a => Expr a -> Expr a -> Expr Bool

eqExpr creates an eq (equal) expression that returns true if the two evaluated input expressions are equal and both don't evaluate to an error.

mkNeExpr :: [AnyExpr] -> Either String AnyExpr

mkNeExpr dynamically creates a ne (not equal) expression, if the two input types are the same.

neExpr :: Eq a => Expr a -> Expr a -> Expr Bool

neExpr creates a ne (not equal) expression that returns true if the two evaluated input expressions are not equal and both don't evaluate to an error.

mkGeExpr :: [AnyExpr] -> Either String AnyExpr

mkGeExpr dynamically creates a ge (greater than or equal) expression, if the two input types are the same.

geExpr :: Ord a => Expr a -> Expr a -> Expr Bool

geExpr creates a ge (greater than or equal) expression that returns true if the first evaluated expression is greater than or equal to the second and both don't evaluate to an error.

mkLeExpr :: [AnyExpr] -> Either String AnyExpr

mkLeExpr dynamically creates a le (less than or equal) expression, if the two input types are the same.

leExpr :: Ord a => Expr a -> Expr a -> Expr Bool

leExpr creates a le (less than or equal) expression that returns true if the first evaluated expression is less than or equal to the second and both don't evaluate to an error.

mkGtExpr :: [AnyExpr] -> Either String AnyExpr

mkGtExpr dynamically creates a gt (greater than) expression, if the two input types are the same.

gtExpr :: Ord a => Expr a -> Expr a -> Expr Bool

gtExpr creates a gt (greater than) expression that returns true if the first evaluated expression is greater than the second and both don't evaluate to an error.

mkLtExpr :: [AnyExpr] -> Either String AnyExpr

mkLtExpr dynamically creates a lt (less than) expression, if the two input types are the same.

ltExpr :: Ord a => Expr a -> Expr a -> Expr Bool

ltExpr creates a lt (less than) expression that returns true if the first evaluated expression is less than the second and both don't evaluate to an error.