Data.Katydid.Relapse.Exprs.Contains

Description

This module contains the Relapse contains expressions.

Documentation

mkContainsExpr :: [AnyExpr] -> Either String AnyExpr

mkContainsExpr dynamically creates a contains expression, if the two input types are:

String and String where the second string is the possible substring.

A List of :Strings, Ints or Uints paired with a String, Int or Uint respectively.

containsStringExpr :: Expr Text -> Expr Text -> Expr Bool

containsStringExpr creates a contains expression that returns true if the second string is a substring of the first.

containsExpr :: Eq a => Expr a -> Expr [a] -> Expr Bool

containsExpr creates a contains expression that returns true if the first argument is an element in the second list argument.