Data.Katydid.Relapse.Exprs.Length

Description

This module contains the Relapse length expressions.

Documentation

mkLengthExpr :: [AnyExpr] -> Either String AnyExpr

mkLengthExpr dynamically creates a length expression, if the single argument is a list, string or bytes.

lengthListExpr :: Expr [a] -> Expr Int

lengthListExpr creates a length expression, that returns the length of a list.

lengthStringExpr :: Expr Text -> Expr Int

lengthStringExpr creates a length expression, that returns the length of a string.

lengthBytesExpr :: Expr ByteString -> Expr Int

lengthBytesExpr creates a length expression, that returns the length of bytes.