Data.Katydid.Relapse.Exprs.Logic

Description

This module contains the Relapse logic expressions: not, and, or.

Documentation

mkNotExpr :: [AnyExpr] -> Either String AnyExpr

mkNotExpr dynamically creates a not expression, if the single argument is a bool expression.

notExpr :: Expr Bool -> Expr Bool

notExpr creates a not expression, that returns true is the argument expression returns an error or false.

mkAndExpr :: [AnyExpr] -> Either String AnyExpr

mkAndExpr dynamically creates an and expression, if the two arguments are both bool expressions.

andExpr :: Expr Bool -> Expr Bool -> Expr Bool

andExpr creates an and expression that returns true if both arguments are true.

mkOrExpr :: [AnyExpr] -> Either String AnyExpr

mkOrExpr dynamically creates an or expression, if the two arguments are both bool expressions.

orExpr :: Expr Bool -> Expr Bool -> Expr Bool

orExpr creates an or expression that returns true if either argument is true.