Data.Katydid.Relapse.Exprs.Type

Description

This module contains the Relapse type expression.

Documentation

mkTypeExpr :: [AnyExpr] -> Either String AnyExpr

mkTypeExpr is used by the parser to create a type expression for the specific input type.

typeExpr :: Expr a -> Expr Bool

typeExpr creates an expression that returns true if the containing expression does not return an error. For example: `(typeExpr varBoolExpr)` will ony return true is the field value is a bool.