Data.Katydid.Relapse.IfExprs

Description

This is an internal relapse module.

It contains multiple implementations of if expressions.

Documentation

data IfExprs

IfExprs is a tree of if expressions, which contains a list of resulting patterns on each of its leaves.

data IfExpr

IfExpr contains a condition and a return pattern for each of the two cases.

newIfExpr :: Expr Bool -> Pattern -> Pattern -> IfExpr

newIfExpr creates an IfExpr.

evalIfExprs :: IfExprs -> Label -> Either String [Pattern]

valIfExprs evaluates a tree of if expressions and returns the resulting patterns or an error.

compileIfExprs :: [IfExpr] -> IfExprs

compileIfExprs compiles a list of if expressions in an IfExprs tree, for efficient evaluation.

data ZippedIfExprs

ZippedIfExprs is a tree of if expressions, but with a zipped pattern list and a zipper on each of the leaves.

zipIfExprs :: IfExprs -> ZippedIfExprs

zipIfExprs compresses an if expression tree's leaves.

evalZippedIfExprs :: ZippedIfExprs -> Label -> Either String ([Pattern], Zipper)

evalZippedIfExprs evaulates a ZippedIfExprs tree and returns the zipped pattern list and zipper from the resulting leaf.