Data.Katydid.Relapse.Parser

Description

This module parses the Relapse Grammar using the Parsec Library.

Parse Grammar

parseGrammar :: String -> Either String Grammar

parseGrammar parses the Relapse Grammar.

parseGrammarWithUDFs :: MkFunc -> String -> Either String Grammar

parseGrammarWithUDFs parses the Relapse Grammar with extra user defined functions.

Internal functions

These functions are exposed for testing purposes.

grammar :: MkFunc -> CharParser () Grammar

For internal testing

pattern :: MkFunc -> CharParser () Pattern

For internal testing

nameExpr :: CharParser () (Expr Bool)

For internal testing

expr :: MkFunc -> CharParser () (Expr Bool)

For internal testing

idLit :: CharParser () String

For internal testing

bytesCastLit :: CharParser () ByteString

For internal testing

stringLit :: CharParser () Text

For internal testing

doubleCastLit :: CharParser () Double

For internal testing

uintCastLit :: CharParser () Word

For internal testing

intLit :: CharParser () Int

For internal testing

ws :: CharParser () ()

For internal testing