Data.Katydid.Relapse.Zip

Description

This is an internal relapse module.

It zips patterns to reduce the state space.

Documentation

data Zipper

Zipper represents compressed indexes that resulted from compressing a list of patterns. This can be used to uncompress a list of bools (nullability of patterns).

zippy :: [Pattern] -> ([Pattern], Zipper)

zippy compresses a list of patterns.

unzipby :: Zipper -> [Bool] -> [Bool]

unzipby uncompresses a list of bools (nullability of patterns).