Expr

Description

This module contains all the Relapse expressions.

It also contains an eval function and a simplfication function for these expressions.

Expressions

data BoolExpr

Constructors

BoolConst Bool

BoolVariable

OrFunc BoolExpr BoolExpr

AndFunc BoolExpr BoolExpr

NotFunc BoolExpr

BoolEqualFunc BoolExpr BoolExpr

DoubleEqualFunc DoubleExpr DoubleExpr

IntEqualFunc IntExpr IntExpr

UintEqualFunc UintExpr UintExpr

StringEqualFunc StringExpr StringExpr

BytesEqualFunc BytesExpr BytesExpr

IntListContainsFunc IntExpr [IntExpr]

StringListContainsFunc StringExpr [StringExpr]

UintListContainsFunc UintExpr [UintExpr]

StringContainsFunc StringExpr StringExpr

BoolListElemFunc [BoolExpr] IntExpr

BytesGreaterOrEqualFunc BytesExpr BytesExpr

DoubleGreaterOrEqualFunc DoubleExpr DoubleExpr

IntGreaterOrEqualFunc IntExpr IntExpr

UintGreaterOrEqualFunc UintExpr UintExpr

BytesGreaterThanFunc BytesExpr BytesExpr

DoubleGreaterThanFunc DoubleExpr DoubleExpr

IntGreaterThanFunc IntExpr IntExpr

UintGreaterThanFunc UintExpr UintExpr

StringHasPrefixFunc StringExpr StringExpr

StringHasSuffixFunc StringExpr StringExpr

BytesLessOrEqualFunc BytesExpr BytesExpr

DoubleLessOrEqualFunc DoubleExpr DoubleExpr

IntLessOrEqualFunc IntExpr IntExpr

UintLessOrEqualFunc UintExpr UintExpr

BytesLessThanFunc BytesExpr BytesExpr

DoubleLessThanFunc DoubleExpr DoubleExpr

IntLessThanFunc IntExpr IntExpr

UintLessThanFunc UintExpr UintExpr

BytesNotEqualFunc BytesExpr BytesExpr

BoolNotEqualFunc BoolExpr BoolExpr

DoubleNotEqualFunc DoubleExpr DoubleExpr

IntNotEqualFunc IntExpr IntExpr

StringNotEqualFunc StringExpr StringExpr

UintNotEqualFunc UintExpr UintExpr

BytesTypeFunc BytesExpr

BoolTypeFunc BoolExpr

DoubleTypeFunc DoubleExpr

IntTypeFunc IntExpr

UintTypeFunc UintExpr

StringTypeFunc StringExpr

RegexFunc StringExpr StringExpr

data DoubleExpr

Constructors

DoubleConst Double

DoubleVariable

DoubleListElemFunc [DoubleExpr] IntExpr

data IntExpr

Constructors

IntConst Int

IntVariable

IntListElemFunc [IntExpr] IntExpr

BytesListLengthFunc [BytesExpr]

BoolListLengthFunc [BoolExpr]

BytesLengthFunc BytesExpr

DoubleListLengthFunc [DoubleExpr]

IntListLengthFunc [IntExpr]

StringListLengthFunc [StringExpr]

UintListLengthFunc [UintExpr]

StringLengthFunc StringExpr

data StringExpr

Constructors

StringConst String

StringVariable

StringListElemFunc [StringExpr] IntExpr

StringToLowerFunc StringExpr

StringToUpperFunc StringExpr

data UintExpr

Constructors

UintConst Int

UintVariable

UintListElemFunc [UintExpr] IntExpr

data BytesExpr

Constructors

BytesConst String

BytesVariable

BytesListElemFunc [BytesExpr] IntExpr

Functions

simplifyBoolExpr :: BoolExpr -> BoolExpr

simplifyBoolExpr returns an equivalent, but simpler version of the input boolean expression.

eval :: BoolExpr -> Label -> Except ValueErr Bool

eval evaluates a boolean expression, given an input label.

Errors

data ValueErr