Exprs.Elem

Description

This module contains the Relapse elem expression.

Documentation

mkElemExpr :: [AnyExpr] -> Either String AnyExpr

mkElemExpr dynamically creates an elem expression, if the first argument is a list and the second an int index.

elemExpr :: Expr [a] -> Expr Int -> Expr a

elemExpr creates an expression that returns an element from the list at the specified index. Trimming this function would cause it to become non generic. It is not necessary to trim each function, since it is just an optimization.