Exprs.Strings

Description

This module contains the Relapse string expressions.

Documentation

mkHasPrefixExpr :: [AnyExpr] -> Either String AnyExpr

mkHasPrefixExpr dynamically creates a hasPrefix expression.

hasPrefixExpr :: Expr Text -> Expr Text -> Expr Bool

hasPrefixExpr creates a hasPrefix expression that returns true if the second is a prefix of the first.

mkHasSuffixExpr :: [AnyExpr] -> Either String AnyExpr

mkHasSuffixExpr dynamically creates a hasSuffix expression.

hasSuffixExpr :: Expr Text -> Expr Text -> Expr Bool

hasSuffixExpr creates a hasSuffix expression that returns true if the second is a suffix of the first.

mkRegexExpr :: [AnyExpr] -> Either String AnyExpr

mkRegexExpr dynamically creates a regex expression.

regexExpr :: Expr Text -> Expr Text -> Expr Bool

regexExpr creates a regex expression that returns true if the first expression matches the second string.

mkToLowerExpr :: [AnyExpr] -> Either String AnyExpr

mkToLowerExpr dynamically creates a toLower expression.

toLowerExpr :: Expr Text -> Expr Text

toLowerExpr creates a toLower expression that converts the input string to a lowercase string.

mkToUpperExpr :: [AnyExpr] -> Either String AnyExpr

mkToUpperExpr dynamically creates a toUpper expression.

toUpperExpr :: Expr Text -> Expr Text

toUpperExpr creates a toUpper expression that converts the input string to an uppercase string.