Parsers

Description

This module describes the abstract tree that can be validated by Relapse.

The JSON and XML parsers both are both versions of this type class.

Documentation

class Tree a where

Tree is the type class that should be implemented by a katydid parser. This is implemented by the Json and XML parser.

Minimal complete definition

getLabel, getChildren

Methods

getLabel :: a -> Label

getChildren :: a -> [a]

data Label

Label is a tagged union of all possible value types that can returned by a katydid parser: String, Int, Uint, Double, Bool and Bytes.

Constructors

String Text

Int Int

Uint Word

Double Double

Bool Bool

Bytes ByteString