Smart

Description

This module describes the smart constructors for Relapse patterns.

Documentation

data Pattern

Pattern recursively describes a Relapse Pattern.

Constructors

Empty

Node

Fields

expr :: Expr Bool

pat :: Pattern

_hash :: Int

Concat

Fields

left :: Pattern

right :: Pattern

_nullable :: Bool

_hash :: Int

Or

Fields

pats :: [Pattern]

_nullable :: Bool

_hash :: Int

And

Fields

pats :: [Pattern]

_nullable :: Bool

_hash :: Int

ZeroOrMore

Fields

pat :: Pattern

_hash :: Int

Reference

Fields

refName :: ValidRef

_nullable :: Bool

_hash :: Int

Not

Fields

pat :: Pattern

_nullable :: Bool

_hash :: Int

ZAny

Contains

Fields

pat :: Pattern

_nullable :: Bool

_hash :: Int

Optional

Fields

pat :: Pattern

_hash :: Int

Interleave

Fields

pats :: [Pattern]

_nullable :: Bool

_hash :: Int

data Grammar

Grammar is a map from reference name to pattern and describes a relapse grammar.

lookupRef :: Grammar -> ValidRef -> Pattern

lookupRef looks up a pattern in the reference map, given a reference name.

compile :: Grammar -> Either String Grammar

compile complies an ast into a smart grammar.

emptyPat :: Pattern

emptyPat is the smart constructor for the empty pattern.

zanyPat :: Pattern

zanyPat is the smart constructor for the zany pattern.

nodePat :: Expr Bool -> Pattern -> Pattern

nodePat is the smart constructor for the node pattern.

orPat :: Pattern -> Pattern -> Pattern

orPat is the smart constructor for the or pattern.

andPat :: Pattern -> Pattern -> Pattern

andPat is the smart constructor for the and pattern.

notPat :: Pattern -> Pattern

notPat is the smart constructor for the not pattern.

concatPat :: Pattern -> Pattern -> Pattern

concatPat is the smart constructor for the concat pattern.

interleavePat :: Pattern -> Pattern -> Pattern

interleavePat is the smart constructor for the interleave pattern.

zeroOrMorePat :: Pattern -> Pattern

zeroOrMorePat is the smart constructor for the zeroOrMore pattern.

optionalPat :: Pattern -> Pattern

optionalPat is the smart constructor for the optional pattern.

containsPat :: Pattern -> Pattern

containsPat is the smart constructor for the contains pattern.

refPat :: Map String Bool -> String -> Either String Pattern

refPat is the smart constructor for the reference pattern.

emptySet :: Pattern

emptySet is the smart constructor for the !(*) pattern.

unescapable :: Pattern -> Bool

unescapable is used for short circuiting. A part of the tree can be skipped if all patterns are unescapable.

nullable :: Pattern -> Bool

nullable returns whether the pattern matches the empty string.

lookupMain :: Grammar -> Pattern

lookupMain retrieves the main pattern from the grammar.