VpaDerive

Description

This module contains a VPA (Visual Pushdown Automaton) implementation of the internal derivative algorithm.

It is intended to be used for explanation purposes.

It shows how out algorithm is effective equivalent to a visual pushdown automaton.

Documentation

derive :: Tree t => Refs -> [t] -> Except String Pattern

derive is the derivative implementation for trees. This implementation makes use of visual pushdown automata.