Expr

Description

This module contains all the functions you need to implement a Relapse expression.

Documentation

data Desc

Desc is the description of a function, especially built to make comparisons of user defined expressions possible.

Constructors

Desc

Fields

_name :: String

_toStr :: String

_hash :: Int

_params :: [Desc]

_hasVar :: Bool

mkDesc :: String -> [Desc] -> Desc

mkDesc makes a description from a function name and a list of the argument's descriptions.

data AnyExpr

AnyExpr is used by the Relapse parser to represent an Expression that can return any type of value, where any is a predefined list of possible types represented by AnyFunc.

Constructors

AnyExpr

Fields

_desc :: Desc

_eval :: AnyFunc

data AnyFunc

AnyFunc is used by the Relapse parser and represents the list all supported types of functions.

Constructors

BoolFunc (Func Bool)

IntFunc (Func Int)

StringFunc (Func Text)

DoubleFunc (Func Double)

UintFunc (Func Word)

BytesFunc (Func ByteString)

BoolsFunc (Func [Bool])

IntsFunc (Func [Int])

StringsFunc (Func [Text])

DoublesFunc (Func [Double])

UintsFunc (Func [Word])

ListOfBytesFunc (Func [ByteString])

data Expr a

Expr represents a user defined expression, which consists of a description for comparisons and an evaluation function.

Constructors

Expr

Fields

desc :: Desc

eval :: Func a

type Func a = Label -> Except String a

Func represents the evaluation function part of a user defined expression. This function takes a label from a tree parser and returns a value or an error string.

params :: Expr a -> [Desc]

params returns the descriptions of the parameters of the user defined expression.

name :: Expr a -> String

name returns the name of the user defined expression.

hasVar :: Expr a -> Bool

hasVar returns whether the expression or any of its children contains a variable expression.

hashWithName :: String -> [Desc] -> Int

hashWithName calculates a hash of the function name and its parameters.

evalConst :: Expr a -> Maybe a

evalConst tries to evaluate a constant expression and either returns the resulting constant value or nothing.

isConst :: Desc -> Bool

isConst returns whether the input description is one of the six possible constant values.

assertArgs1 :: String -> [AnyExpr] -> Except String AnyExpr

assertArgs1 asserts that the list of arguments is only one argument and returns the argument or an error message containing the function name that was passed in as an argument to assertArgs1.

assertArgs2 :: String -> [AnyExpr] -> Except String (AnyExpr, AnyExpr)

assertArgs2 asserts that the list of arguments is only two arguments and returns the two arguments or an error message containing the function name that was passed in as an argument to assertArgs2.

mkBoolExpr :: Expr Bool -> AnyExpr

mkBoolExpr generalises a bool expression to any expression.

mkIntExpr :: Expr Int -> AnyExpr

mkIntExpr generalises an int expression to any expression.

mkStringExpr :: Expr Text -> AnyExpr

mkStringExpr generalises a string expression to any expression.

mkDoubleExpr :: Expr Double -> AnyExpr

mkDoubleExpr generalises a double expression to any expression.

mkBytesExpr :: Expr ByteString -> AnyExpr

mkBytesExpr generalises a bytes expression to any expression.

mkUintExpr :: Expr Word -> AnyExpr

mkUintExpr generalises a uint expression to any expression.

assertBool :: AnyExpr -> Except String (Expr Bool)

assertBool asserts that any expression is actually a bool expression.

assertInt :: AnyExpr -> Except String (Expr Int)

assertInt asserts that any expression is actually an int expression.

assertString :: AnyExpr -> Except String (Expr Text)

assertString asserts that any expression is actually a string expression.

assertDouble :: AnyExpr -> Except String (Expr Double)

assertDouble asserts that any expression is actually a double expression.

assertBytes :: AnyExpr -> Except String (Expr ByteString)

assertBytes asserts that any expression is actually a bytes expression.

assertUint :: AnyExpr -> Except String (Expr Word)

assertUint asserts that any expression is actually a uint expression.

boolExpr :: Bool -> Expr Bool

boolExpr creates a constant bool expression from a input value.

intExpr :: Int -> Expr Int

intExpr creates a constant int expression from a input value.

stringExpr :: Text -> Expr Text

stringExpr creates a constant string expression from a input value.

doubleExpr :: Double -> Expr Double

doubleExpr creates a constant double expression from a input value.

bytesExpr :: ByteString -> Expr ByteString

bytesExpr creates a constant bytes expression from a input value.

uintExpr :: Word -> Expr Word

uintExpr creates a constant uint expression from a input value.

trimBool :: Expr Bool -> Expr Bool

trimBool tries to reduce an expression to a single constant expression, if it does not contain a variable.

trimInt :: Expr Int -> Expr Int

trimInt tries to reduce an expression to a single constant expression, if it does not contain a variable.

trimString :: Expr Text -> Expr Text

trimString tries to reduce an expression to a single constant expression, if it does not contain a variable.

trimDouble :: Expr Double -> Expr Double

trimDouble tries to reduce an expression to a single constant expression, if it does not contain a variable.

trimBytes :: Expr ByteString -> Expr ByteString

trimBytes tries to reduce an expression to a single constant expression, if it does not contain a variable.

trimUint :: Expr Word -> Expr Word

trimUint tries to reduce an expression to a single constant expression, if it does not contain a variable.

mkBoolsExpr :: Expr [Bool] -> AnyExpr

mkBoolsExpr generalises a list of bools expression to any expression.

mkIntsExpr :: Expr [Int] -> AnyExpr

mkIntsExpr generalises a list of ints expression to any expression.

mkStringsExpr :: Expr [Text] -> AnyExpr

mkStringsExpr generalises a list of strings expression to any expression.

mkDoublesExpr :: Expr [Double] -> AnyExpr

mkDoublesExpr generalises a list of doubles expression to any expression.

mkListOfBytesExpr :: Expr [ByteString] -> AnyExpr

mkListOfBytesExpr generalises a list of bytes expression to any expression.

mkUintsExpr :: Expr [Word] -> AnyExpr

mkUintsExpr generalises a list of uints expression to any expression.

assertBools :: AnyExpr -> Except String (Expr [Bool])

assertBools asserts that any expression is actually a list of bools expression.

assertInts :: AnyExpr -> Except String (Expr [Int])

assertInts asserts that any expression is actually a list of ints expression.

assertStrings :: AnyExpr -> Except String (Expr [Text])

assertStrings asserts that any expression is actually a list of strings expression.

assertDoubles :: AnyExpr -> Except String (Expr [Double])

assertDoubles asserts that any expression is actually a list of doubles expression.

assertListOfBytes :: AnyExpr -> Except String (Expr [ByteString])

assertListOfBytes asserts that any expression is actually a list of bytes expression.

assertUints :: AnyExpr -> Except String (Expr [Word])

assertUints asserts that any expression is actually a list of uints expression.

boolsExpr :: [Expr Bool] -> Expr [Bool]

boolsExpr sequences a list of expressions that each return a bool, to a single expression that returns a list of bools.

intsExpr :: [Expr Int] -> Expr [Int]

intsExpr sequences a list of expressions that each return an int, to a single expression that returns a list of ints.

stringsExpr :: [Expr Text] -> Expr [Text]

stringsExpr sequences a list of expressions that each return a string, to a single expression that returns a list of strings.

doublesExpr :: [Expr Double] -> Expr [Double]

doublesExpr sequences a list of expressions that each return a double, to a single expression that returns a list of doubles.

listOfBytesExpr :: [Expr ByteString] -> Expr [ByteString]

listOfBytesExpr sequences a list of expressions that each return bytes, to a single expression that returns a list of bytes.

uintsExpr :: [Expr Word] -> Expr [Word]

uintsExpr sequences a list of expressions that each return a uint, to a single expression that returns a list of uints.