Exprs

Description

This module contains the standard library of expressions, used by the Relapse parser.

Documentation

mkBuiltIn :: String -> AnyExpr -> Either String AnyExpr

mkBuiltIn parsers a builtin function to a relapse expression.

mkExpr :: String -> [AnyExpr] -> Either String AnyExpr

mkExpr is a grouping of all the standard library functions as one MkFunc.

type MkFunc = String -> [AnyExpr] -> Either String AnyExpr

MkFunc is used by the parser to create a function from a name and arguments.

stdOnly :: String -> [AnyExpr] -> Either String AnyExpr

stdOnly contains no functions, which means that when it is combined (in Relapse parser) with mkExpr the parser will have access to only the standard library.